test(bindCallback): replaced test case under bindCallback test
